Make the layer build on the four platforms the node's CI covers
Run #406 failed on macos-arm64, macos-x86_64, linux-arm64 and windows-x86_64.
None of it is environment: the layer was only ever built on x86_64 Linux, and
the node's CI builds the whole workspace on five platforms. Three distinct
portability defects, all in layer crates.

btc-vault/src/seal.rs: five items — the key-store magic, version, envelope
length and the encode/decode pair — sat outside the file's pervasive
target_os = "linux" gating while their only consumers sat inside it. Off Linux
they are unreachable and -D dead-code rejects them. The gate is placed on the
definitions rather than the lint silenced, so they exist exactly where used.

dom-scriptless-store runtime/linux.rs: link_count is declared u64 while
stat.st_nlink follows the architecture — libc 0.2.186 defines nlink_t as u64
on x86_64-linux and u32 on aarch64-linux. Converted with u64::from, and the
useless-conversion allow is scoped to that one field with the reason written,
because the conversion is real on one arch and reflexive on the other.

relay:  was gated on its feature but not on the platform,
and that feature is default-on. Its substance is Linux-only — retained
directory capabilities, rustix::process (which upstream gates
cfg(not(windows))), Unix permission modes — so a --workspace build on Windows
reached thirteen errors' worth of code that cannot exist there. Module, its
 companion which imports from it, and the production_durable test are
gated to Linux.

Verified on x86_64 Linux with the same RUSTFLAGS=-D warnings the CI build step
uses — which my local gate did NOT set, and that gap is why dead_code and
unused_imports reached CI as errors. The other platforms cannot be built here:
cross-checking dies in the C build scripts of secp256k1-sys and
libsqlite3-sys, so CI is the oracle for those.

+++ crates/dom-scriptless-store/src/runtime/linux.rs
@@ -206,7 +206,13 @@ impl NodeIdentity {
             device: stat.st_dev,
             inode: stat.st_ino,
             mode: stat.st_mode,
-            link_count: stat.st_nlink,
+            // `st_nlink` follows the architecture: libc defines `nlink_t` as
+            // u64 on x86_64-linux and u32 on aarch64-linux. The conversion is
+            // real on aarch64 and reflexive on x86_64, where clippy calls it
+            // useless — so the allow is scoped to this one field rather than
+            // the field's type being pinned to whatever the build host uses.
+            #[allow(clippy::useless_conversion)]
+            link_count: u64::from(stat.st_nlink),
             owner: stat.st_uid,
             node_type: expected_type,
         })
